Refactoring and cleaning the code.
I refactored a code, functions from handlers.js I put straight in the
components, in order to fill choose table short info I made a context
with global state with a dict in it. In createBookingReqeuests.js and
booking.js I set the value of a dict and in chooseTable.js I use context
and get the values and use it. Instead of formatting and parsing the
dates with flatpickr I do it with date-fns. In createBookingRequests I
made more components (Divided formik into different components). Instead
of reactDom.render started to use root. In createBookingRequest you now
see chosen table hover (hover around div with table info) from the start.