view booking detail + booking delete
I added booking detail page, from that page you can also delete that
booking. If user makes request manually (dont know why) Drf checks if
booking he wants to delete or view is created by him or not.
Also I made it so if chosen table from the localStorage is no
available and not present it sets chosen table to last table of a list.